+++ packages/core/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @better-webhook/core
 
+## 0.4.0
+
+### Minor Changes
+
+- Minor refactor of the core module to improve readability, maintainability, and performance.
+
+## 0.3.0
+
+### Minor Changes
+
+- Adds request context and updates docs
+
 ## 0.2.0
 
 ### Minor Changes