+++ packages/express/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @better-webhook/express
 
+## 0.5.0
+
+### Minor Changes
+
+- Minor refactor of the Express module to improve readability, maintainability, and performance.
+
+### Patch Changes
+
+- Updated dependencies
+  - @better-webhook/core@0.4.0
+
+## 0.4.0
+
+### Minor Changes
+
+- Adds request context and updated documentation
+
+## 0.3.0
+
+### Minor Changes
+
+- Adds request context and updates docs
+
+### Patch Changes
+
+- Updated dependencies
+  - @better-webhook/core@0.3.0
+
 ## 0.2.0
 
 ### Minor Changes