+++ packages/nextjs/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @better-webhook/nextjs
 
+## 0.4.0
+
+### Minor Changes
+
+- Minor refactor of the NextJS module to improve readability, maintainability, and performance.
+
+### Patch Changes
+
+- Updated dependencies
+  - @better-webhook/core@0.4.0
+
+## 0.3.0
+
+### Minor Changes
+
+- Adds request context and updates docs
+
+### Patch Changes
+
+- Updated dependencies
+  - @better-webhook/core@0.3.0
+
 ## 0.2.0
 
 ### Minor Changes